% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{add_emas}
\alias{add_emas}
\title{Compute and Add Exponential Moving Averages (EMAs) to Data Frame}
\usage{
add_emas(df, periods = c(20, 50, 100, 200, 500, 1000))
}
\arguments{
\item{df}{A data frame to which the EMAs are to be appended. This data frame should have at least the columns:
'symbol' (indicating the stock ticker), 'adjusted' (representing the adjusted closing prices), and 'n'
(the count of observations for each ticker).}

\item{periods}{A numeric vector of non-negative integers representing the periods for which EMAs are to be
computed. For instance, \code{c(5, 20, 50)} would compute EMAs for 5-day, 20-day, and 50-day periods.}
}
\value{
A data frame that includes the original data and additional columns for each EMA computed for the specified
periods. The newly added EMA columns are named in the format: "ema_" followed by the period number (e.g.,
"ema_5" for a 5-day EMA). Each EMA is computed independently for each unique 'symbol'. If a 'symbol'
doesn't have sufficient data for a particular period, the corresponding EMA column will have \code{NA} values.
}
\description{
The \code{add_emas} function enriches a given financial time series data frame by computing and adding columns
for Exponential Moving Averages (EMAs) for specified periods. EMAs provide insight into trends in financial data
by giving more weight to recent data points.
}
\details{
For each unique stock ticker ('symbol') in the data frame, the function computes EMAs only if the number of
observations for that ticker is greater than the specified EMA period. If not, it leaves the EMA value as \code{NA}
for that ticker and period combination.
}
\examples{
\dontrun{
data <- data.frame(
  symbol = rep("AAPL", 5),
  adjusted = c(145, 146, 148, 149, 150),
  n = c(5, 5, 5, 5, 5)
)
enhanced_data <- add_emas(data, c(3, 5))
}

}
