% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{add_pct_distances_from_emas}
\alias{add_pct_distances_from_emas}
\title{Add Percentage Distances from Exponential Moving Averages (EMAs)}
\usage{
add_pct_distances_from_emas(df, periods = c(20, 50, 100, 200, 500, 1000))
}
\arguments{
\item{df}{A data frame that contains at least the columns 'adjusted' and 'date'. It should also have the EMA columns
for the given periods in the format "ema_" followed by the period number (e.g., "ema_5" for a 5-day EMA).}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the metrics are to be
computed. Default periods are c(20, 50, 100, 200, 500, 1000).}
}
\value{
A data frame with original data and additional columns representing the percentage distance from the EMAs
for the specified periods.
}
\description{
The \code{add_pct_distances_from_emas} function enhances a financial time series data frame by calculating the percentage
distances from given EMAs.
}
\details{
The function computes the percentage distance from each EMA as \code{(adjusted - EMA) / EMA}.
}
