% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate Correlation Coefficients}
\usage{
calculate_correlations(df, prefix = "")
}
\arguments{
\item{df}{A data frame that contains at least the columns starting with 'z_score_' and 'return_'.}

\item{prefix}{An optional prefix to be added to 'z_score_' if the z-scores columns have additional prefixes. Default is an empty string.}
}
\value{
A data frame containing the computed correlation coefficients for all combinations of z-scores and returns.
}
\description{
The function \code{calculate_correlations} computes the correlation coefficients between the z-scores and returns
for all combinations of 'z_score_' and 'return_' columns in the given data frame.
}
\examples{
\dontrun{
  correlations <- calculate_correlations(df)
}

}
