% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{calculate_linear_coefficients}
\alias{calculate_linear_coefficients}
\title{Calculate Linear Regression and Summarize Coefficients}
\usage{
calculate_linear_coefficients(
  df,
  periods = c(20, 50, 100, 200, 500, 1000),
  return_periods = c(20, 50, 100, 200, 500, 1000),
  prefix = "",
  prefix_2 = NULL
)
}
\arguments{
\item{df}{A data frame that must contain rounded columns named 'z_score_' and columns named 'return_'
followed by corresponding period numbers. E.g., 'z_score_20', 'return_20', etc.}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the
linear regressions and summaries are to be computed. Default periods are
c(20, 50, 100, 200, 500, 1000).}

\item{return_periods}{A numeric vector of non-negative integers representing the periods over which
returns are to be computed. Vector must be same length as \code{periods}. Default periods are
c(20, 50, 100, 200, 500, 1000).}
}
\value{
A data frame containing the computed summaries and linear regressions for the specified periods,
including mean return, standard deviation of return, sample size, duration, regression,
mean daily return, standard deviation of daily return, and daily regression.
}
\description{
This function calculates linear regressions for pairs of return and z-score variables
for specified periods in a given data frame. It returns a data frame with the mean and
modeled return for each z-score stratum, including additional summary statistics.
}
\examples{
\dontrun{
  plotdata <- calculate_linear_coefficients(df, periods = c(20, 50, 100, 200, 500, 1000))
}
}
