% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{calculate_returns}
\alias{calculate_returns}
\title{Calculate Stock Price Returns}
\usage{
calculate_returns(df, periods = c(20, 50, 100, 200, 500, 1000))
}
\arguments{
\item{df}{A data frame that contains at least the columns 'adjusted', 'date', and 'symbol'.}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the returns are to be computed.
Default periods are c(20, 50, 100, 200, 500, 1000).}
}
\value{
A data frame with the original data and additional columns containing the computed returns for the specified periods,
excluding symbols with extremely aberrant 20-day returns.
}
\description{
The function \code{calculate_returns} computes the stock price returns for specified periods. It takes a data frame
containing stock price data and returns a new data frame with additional columns representing the returns
for the specified periods. Symbols with extremely aberrant 20-day returns are excluded from the result.
}
\examples{
\dontrun{
  df_with_returns <- calculate_returns(df)
}

}
