% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{calculate_rolling_beta}
\alias{calculate_rolling_beta}
\title{Calculate Rolling Beta for Stocks}
\usage{
calculate_rolling_beta(
  df,
  index = tq_get("^GSPC"),
  window_size = 2,
  window_in_years = TRUE,
  partial = TRUE
)
}
\arguments{
\item{df}{A data frame containing at least the columns 'symbol', 'date', and 'adjusted', representing the daily data for the stocks.}

\item{index}{A data frame containing price history for a benchmark index (e.g., "^GSPC" for the S&P 500 index). Default is S&P 500 price history fetched with tidyquant.}

\item{window_size}{Numeric value specifying the window size for the rolling calculation. Default is 2.}

\item{window_in_years}{Logical value indicating whether the window_size is provided in years (TRUE) or in trading days (FALSE). Default is TRUE.}

\item{partial}{Logical value indicating whether to calculate beta with partial data at the edges of the time series if full window size is not available. Default is TRUE.}
}
\value{
A data frame with the original data (from \code{df}) and an additional 'rolling_beta' column containing the computed rolling beta values.
}
\description{
The function \code{calculate_rolling_beta} computes the rolling beta for a given set of stocks relative to a specified index
over a specified window size. It takes a data frame containing daily adjusted prices for the stocks and a corresponding benchmark index data frame
and returns a data frame with the rolling beta values.
}
\examples{
\dontrun{
  stocks_with_beta <- calculate_rolling_beta(df, window_size = 5)
}
}
