% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{classify_investment_thesis}
\alias{classify_investment_thesis}
\title{Investment Thesis Classification Function}
\usage{
classify_investment_thesis(df)
}
\arguments{
\item{df}{A data frame that houses the data requiring analysis. It should include columns that start with
"implied_daily_return_" representing daily investment returns and those that start with "moving_average_"
indicating the moving averages of returns.}
}
\value{
A data frame enriched with an additional column named 'type' that classifies each row's investment thesis
based on the analyzed data. This data frame is returned in descending order based on the column
'average_expected_return'.
}
\description{
The \code{classify_investment_thesis} function takes a data frame (\code{df}) that contains data related to investment returns.
Based on the means of short and long periods for these returns, the function determines an investment thesis classification
which is either 'short-term mean reversion', 'short-term momentum', 'long-term mean reversion', or 'long-term momentum'.
This classification is crucial for investment strategizing and understanding the nature of returns over different periods.
}
\details{
The function processes the provided data using the following key steps:
\enumerate{
\item Defines mean calculation functions for short and long periods both for direct returns and moving averages.
\item Calculates the 'average_expected_return' for each row in the data frame.
\item Assigns a type of investment thesis based on a combination of the mean values of short and long periods
and their corresponding moving averages.
\item Returns the modified data frame with the investment thesis classification and the data sorted in descending order
of 'average_expected_return'.
}
}
\examples{
\dontrun{
classified_data <- classify_investment_thesis(investment_df)
}

}
