% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{compute_linear_models_by_z_score_stratum}
\alias{compute_linear_models_by_z_score_stratum}
\title{Compute Linear Models by Different Z-Score Strata}
\usage{
compute_linear_models_by_z_score_stratum(
  df,
  periods = c(20, 50, 100, 200, 500, 1000),
  return_periods = c(20, 50, 100, 200, 500, 1000),
  filename_prefix = ""
)
}
\arguments{
\item{df}{A data frame containing the necessary variables for the linear model computation.}

\item{filename_prefix}{A string used to prepend to the filenames of the saved results. Default is an empty string.}
}
\value{
NULL. Results are saved to files within the ./data/ directory.
}
\description{
This function computes linear models for each z-score stratum, each 20-day-moving-average-of-z-score stratum,
and double-stratifying by 20-day moving average of z-score and the current z-score's difference from it.
It uses the \code{calculate_linear_coefficients} function for these computations and saves the results to files.
}
\examples{
\dontrun{
  compute_linear_models_by_z_score_stratum(df, filename_prefix = "sp500_")
}
}
\seealso{
\code{\link{calculate_linear_coefficients}} for the function used to calculate linear coefficients.
}
