% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{compute_moving_averages}
\alias{compute_moving_averages}
\title{Compute Moving Averages of Z-Scores and Differences Between Z-Scores and Their Moving Averages}
\usage{
compute_moving_averages(
  df,
  periods = c(20, 50, 100, 200, 500, 1000),
  ma_period = 20,
  use_only_past_data = FALSE
)
}
\arguments{
\item{df}{A data frame with the z-scores and percentage distances from EMAs.}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the metrics are to be
computed. Default periods are c(20, 50, 100, 200, 500, 1000).}

\item{ma_period}{A single non-negative integer that sets the period for the simple moving average computation. Default is 20.}

\item{use_only_past_data}{A boolean. If TRUE, the z-scores will be calculated using only prior data for each row. Default is FALSE.}
}
\value{
A data frame with the computed moving averages of z-scores and the differences between z-scores and their moving averages.
}
\description{
This function calculates the moving averages of z-scores and the differences between the z-scores
and their moving averages. If the \code{use_only_past_data} parameter is TRUE, it computes the z-scores
using only prior data for each row.
}
