% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{convert_distances_to_zscores}
\alias{convert_distances_to_zscores}
\title{Convert Percentage Distances to Z-Scores}
\usage{
convert_distances_to_zscores(
  df,
  periods = c(20, 50, 100, 200, 500, 1000),
  use_only_past_data = FALSE,
  window_size = NULL,
  window_in_years = TRUE
)
}
\arguments{
\item{df}{A data frame that contains at least the columns 'adjusted', 'date', and percentage distances from EMAs for the specified periods.}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the z-scores are to be computed. Default periods are c(20, 50, 100, 200, 500, 1000).}

\item{use_only_past_data}{A boolean indicating whether to use only past data for z-score calculations. If TRUE, z-scores are computed using rolling calculations. Default is FALSE.}

\item{window_size}{A numeric value specifying the rolling window size for z-score calculations. Can be given in years or rows. If NULL, the full history of data is used when \code{use_only_past_data} is TRUE. Default is NULL.}

\item{window_in_years}{A boolean indicating whether the \code{window_size} is given in years. If TRUE, \code{window_size} is converted to rows using 252 trading days in a year. Default is TRUE.}
}
\value{
A data frame with the original data and additional columns containing the computed z-scores for the specified periods.
}
\description{
The function \code{convert_distances_to_zscores} computes the z-scores from percentage distances from given Exponential Moving Averages (EMAs).
You can specify whether to use only past data for calculations, and the window size for the rolling calculations.
}
\examples{
\dontrun{
  data <- convert_distances_to_zscores(data, periods = c(50, 100), use_only_past_data = TRUE, window_size = 2, window_in_years = TRUE)
}

}
