% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{get_prices}
\alias{get_prices}
\title{Fetch Stock Price History For Given Tickers}
\usage{
get_prices(tickers, include_today = FALSE)
}
\arguments{
\item{tickers}{A character vector specifying the stock tickers you're interested in.}

\item{include_today}{Logical. If TRUE, the function will also fetch and include stock prices
for the current trading day. Default is FALSE.}
}
\value{
A data frame containing the historical stock prices. Each row represents a day of trading for a
particular stock. The main columns are \code{symbol} (stock ticker), \code{date} (the trading date),
and \code{adjusted} (the adjusted closing price).
}
\description{
This function retrieves stock price history from various sources for a list of provided tickers.
Specifically, it makes use of the \code{tidyquant} and \code{quantmod} packages to fetch data.
Optionally, it can also retrieve the most current trading day's stock price.
}
\details{
Please note: If you're using the Alpha Vantage source (\code{src = "av"}) for fetching today's prices,
ensure you have set your Alpha Vantage API key using \code{Sys.setenv(ALPHAVANTAGE_API_KEY = "your_key")}.
}
\examples{
\dontrun{
Sys.setenv(ALPHAVANTAGE_API_KEY = "your_key_here")
get_prices(c("AAPL", "MSFT"), include_today = TRUE)
}

}
