% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{plot_results}
\alias{plot_results}
\title{Plot Returns and Error Bars for Z-Score or Moving Average Thereof}
\usage{
plot_results(
  plotdata,
  stratum = "z_score_stratum",
  asset = "an S&P 500 stock",
  periods = c(20, 50, 100, 200, 500, 1000)
)
}
\arguments{
\item{plotdata}{A data frame containing the required plot data, including mean daily return and standard deviation of daily return.}

\item{stratum}{A string representing the column used for stratification. Default is "z_score_stratum".}

\item{asset}{A string representing the asset name. Default is "an S&P 500 stock".}

\item{periods}{A numeric vector of integers representing the periods for which the plot is to be generated. Default periods are c(20, 50, 100, 200, 500, 1000).}
}
\value{
A ggplot object representing the plotted data.
}
\description{
This function plots the mean daily return for investing in an asset, such as an S&P 500 stock,
by the price's z-score difference from EMA, or the trailing 20-day moving average thereof, or the difference between
these two. Error bars are also added to represent the standard deviation of daily returns.
}
\examples{
\dontrun{
  plot_results(plotdata, stratum = "z_score_stratum", asset = "an S&P 500 stock", periods = c(20, 50, 100, 200, 500, 1000))
}
}
