% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{plot_results_ds}
\alias{plot_results_ds}
\title{Plot Returns for Double Stratified Data}
\usage{
plot_results_ds(
  plotdata,
  stratum1 = "moving_average_z_score_stratum",
  stratum2 = "diff_z_score_stratum",
  asset = "an S&P 500 stock",
  period = 20
)
}
\arguments{
\item{plotdata}{A data frame containing the required plot data, including mean daily return and standard deviation of daily return.}

\item{stratum1}{A string representing the first stratification column. Default is "moving_average_z_score_stratum".}

\item{stratum2}{A string representing the second stratification column. Default is "diff_z_score_stratum".}

\item{asset}{A string representing the asset name. Default is "an S&P 500 stock".}

\item{period}{An integer representing the EMA period for the analysis. Default is 20.}
}
\value{
A ggplot object representing the plotted data.
}
\description{
This function plots the mean daily return for investing in an asset (e.g., an S&P 500 stock) by the difference
between the z-score relationship to a specified exponential moving average (EMA) period and the trailing 20-day
moving average thereof. The plot is double stratified according to the parameters specified.
}
\examples{
\dontrun{
  plotresults_ds(plotdata, stratum1 = "moving_average_z_score_stratum", stratum2 = "diff_z_score_stratum", asset = "an S&P 500 stock", period = 20)
}
}
