% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{prep_data_for_analysis}
\alias{prep_data_for_analysis}
\title{Data Preparation for Analysis}
\usage{
prep_data_for_analysis(
  tickers,
  periods = c(20, 50, 100, 200, 500, 1000),
  return_periods = c(20, 50, 100, 200, 500, 1000),
  ma_period_to_use = 20,
  window_size_to_use_in_years = 2,
  benchmark_index_ticker = "^GSPC",
  include_today = FALSE
)
}
\description{
The \code{prep_data_for_analysis} function takes a vector of ticker symbols, retrieves their historical price data,
and prepares the data for further analysis.
}
