% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{prep_data_for_forecasting}
\alias{prep_data_for_forecasting}
\title{Data Preparation for Return Forecasting}
\usage{
prep_data_for_forecasting(
  tickers,
  periods = c(20, 50, 100, 200, 500, 1000),
  include_today = FALSE
)
}
\arguments{
\item{tickers}{A character vector of ticker symbols for which the data needs to be prepared.
Example: \code{c("AAPL", "GOOGL", "MSFT")}.}

\item{periods}{A numeric vector specifying the periods (in days) for which EMAs should be computed.
For instance, \code{c(5, 20, 50)} would request the 5-day, 20-day, and 50-day EMAs.}

\item{include_today}{A logical value indicating whether today's price data should be included in the analysis.
If \code{TRUE}, the most recent price data will be appended. Defaults to \code{FALSE}.}
}
\value{
A data frame prepared for analysis, encompassing cleaned price data, EMAs, z-score distances
from the EMAs, and a column indicating the number of observations for each ticker symbol.
}
\description{
The \code{prep_data_for_forecasting} function takes a vector of ticker symbols, retrieves their historical price data,
and prepares the data for further analysis. The data preparation encompasses several steps:
\enumerate{
\item Retrieval of historical price data for the provided tickers using the \code{get_prices} function.
\item Cleaning of the data, which involves filtering out NAs, removing symbols with insufficient observations,
and appending a column with the count of observations for each symbol using the \code{clean_data} function.
\item Computation and addition of Exponential Moving Averages (EMAs) for specified periods via the \code{add_emas} function.
\item Calculation and inclusion of z-score distances from the computed EMAs with the help of the
\code{add_pct_distances_from_emas} function.
}
}
\examples{
\dontrun{
prepared_data <- prep_data_for_forecasting(c("AAPL", "GOOGL"), c(5, 20))
}

}
