% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{round_metrics}
\alias{round_metrics}
\title{Round Z-Scores, Moving Averages, and Their Differences}
\usage{
round_metrics(df, periods = c(20, 50, 100, 200, 500, 1000))
}
\arguments{
\item{df}{A data frame with z-scores, moving averages of z-scores, and the differences between z-scores and their moving averages.}

\item{periods}{A numeric vector of non-negative integers representing the periods for which the metrics are to be
computed. Default periods are c(20, 50, 100, 200, 500, 1000).}
}
\value{
A data frame with rounded z-scores, moving averages, and their differences.
}
\description{
This function rounds z-scores, moving averages of z-scores, and the differences between z-scores and their moving averages
to the nearest integer.
}
