% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{time_series_split}
\alias{time_series_split}
\title{Time Series Splitting Function for Training and Testing}
\usage{
time_series_split(df, split_date, n_of_rows_to_omit)
}
\arguments{
\item{df}{A data frame containing at least the columns 'symbol', 'date', and 'adjusted', representing the daily data for the stocks.}

\item{split_date}{A Date object representing the cut-off date for splitting the data into training and testing sets.}

\item{n_of_rows_to_omit}{An integer specifying the number of rows to omit at the end of the training set for each symbol.}
}
\value{
A list containing two data frames: 'train' and 'test'.
}
\description{
The function \code{time_series_split} splits the time series data into training and testing sets by date, grouped by symbols.
It allows to specify the number of rows to omit from the training set to prevent overlap between test and train for the outcome variable.
}
\examples{
\dontrun{
  split_data <- time_series_split(df, as.Date("2020-01-01"), n_of_rows_to_omit = 5)
}
}
