% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{trade_potential}
\alias{trade_potential}
\title{Calculate Trade Potential}
\usage{
trade_potential(plotdata, strategy)
}
\arguments{
\item{plotdata}{A data frame containing the required data, including mean daily return, duration, and number of trades (n).}

\item{strategy}{A string indicating the strategy for tidy result grouping and comparison}
}
\value{
A data frame with the summarized potential return and the number of trades for each duration.
}
\description{
This function filters the data for trades with a mean daily return greater than 0.0008, then calculates the
potential return for different durations. It returns a summarized view with the overall potential return
and the number of trades for each duration.
}
\examples{
\dontrun{
  potential <- trade_potential(plotdata)
  print(potential)
}
}
