% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{analyze_data}
\alias{analyze_data}
\title{Financial Data Analysis Function}
\usage{
analyze_data(
  df,
  plotdata,
  stratum1 = "z_score_stratum",
  stratum2 = NULL,
  periods = c(20, 50, 100, 200, 500, 1000)
)
}
\arguments{
\item{df}{A data frame containing the financial data that needs to be analyzed.
This data frame should have columns corresponding to the stratum (or stratums) specified.}

\item{plotdata}{A dataframe containing pre-calculated plot data.
This data is crucial for the expected returns calculation.}

\item{stratum1}{A character string that designates the primary stratum based on which expected returns are computed.}

\item{stratum2}{An optional character string indicating a secondary stratum to consider when calculating expected returns.
By default, it is set to \code{NULL}, meaning only \code{stratum1} is considered.}
}
\value{
A data frame containing the financial data from \code{df} supplemented with the average expected returns
for each observation, arranged in descending order of the average expected return.
}
\description{
The \code{analyze_data} function performs an analysis on the provided financial data frame (\code{df}) based on pre-defined
stratum or stratums in the pre-calculated plotdata.
It calculates the expected returns for the designated stratum (or stratums) and computes their averages.
If a secondary stratum (\code{stratum2}) is provided, it factors that into the expected returns calculation.
Finally, the analyzed data is returned in a data frame that's sorted by the average expected return in descending order.
}
\details{
The function processes the financial data using the following steps:
\enumerate{
\item It iteratively goes through each row of the \code{df} data frame.
\item For each iteration (or row), it matches the stratum values with the precalculated plot data to get the expected return.
\item If \code{stratum2} is provided, the function performs a dual-stratum matching.
\item It then computes the average of all matched expected returns for the row.
\item The resultant data frame contains the original data along with the average expected returns.
}
}
\examples{
\dontrun{
analyzed_data <- analyze_data(financial_df, plotdata, "stratum_name")
}

}
